// src/app/modelo/aula-curso-dto.ts
export class AulaCursoDTO {
    idFacultadCurso: string = "";
    idPrograma: string = "";
    idCurso: string = "";
    idGrupo: number = 0;
    idHorario: string = "";
    idFacultadAula: string = "";
    idAula: string = "";
    diaSemana: number = 0;
    horaInicio: number = 0;
    horaFin: number = 0;
}